chore: move lambda_function.py and requirements.txt under imgupload/ and update tests import
